edit commit

+++ Programming_I/Algorithm/Enshu0101.java
@@ -1,6 +1,6 @@
 public class Enshu0101{
 	public static void main(String[] args){
-		double nuki=-1,komi;
+		int nuki=-1,komi;
 		java.util.Scanner stdScanner=new java.util.Scanner(System.in);
 		
 		while(nuki<0){
@@ -9,7 +9,7 @@ public class Enshu0101{
 			if(nuki<0)
 				System.out.println("0以上の値を入力してください");
 		}
-		komi=nuki*1.10;
+		komi=(int)(nuki*1.10);
 		
 		System.out.printf("税込み価格 : %.0f\n",komi);
 		stdScanner.close();
